Refactor useAuth hook
In order to streamline the use of useAuth hook, changed default export
to the context consumer instead of the hook function itself. This way
importing useAuth in a component will allow direct interaction with
the context object.

As a side effect it was necessary to move the AuthProvider component
inside the hook file itself.