Fix infinite loop in useFetch
Two properties were using empty objects as default values, which ended
up being redeclared every time the hook re-rendered. That in turn would
trigger the useEffect causing an infinite loop. The default empty
objects are now declared as variables outside the scope of the function
preventing that from happening.

The queryString obj also had to be memoized with useMemo.