Fix NotificationDisplay multiple events bug
setTimeout callback inside NotificationDisplay was using an old state
value to update eventQueue after a notification was shown. This meant
that any new events added between the notification being displayed and
the setTimeout callback running would be ignored by the component.

Changed the setState call to take a callback instead, allowing the
function to use the up-to-date state value whenever setTimeout executes.