Fix new messages not showing in empty channels
After refactoring the api to send a 404 response when no chat messages
have been found, useChatMessages would no longer create an entry in
messageStore for that text channel. This meant that any new messages
coming in over websockets had no message store to be added to and would
be ignored until the user refreshed the app, causing a successful fetch
from the api.

Now, regardless of the response from the api a message array gets
created inside the messageStore, fixing the issue.