Refactor useUserList
useUserList has been refactored to store a mapping between servers and
its members, instead of channel and its users. This reflects the
changes to API that now assign an extra SocketIo room to the entire
server, not just specific channels. The hook no longer needs to make
multiple fetches for user status, one for each channel. Now it can get
get that data for all the server members, and any context consumers
can extract a subset of the user list that represents members of one
specific channel.