Export socket event inside useSocket
Refactored useSocket to re-export the socket element. The reason for
doing this is the massive amount of redundant code written for every
single type of socket event. The previous implementation involved
two separate message emitters - one included with SocketIo client
package and another added manually, as well as a whole hook
useSocketEvents that stored arrays of all the socket events. This was
incredibly inefficient and tedious to work with. Re-exporting the
socket element through useSocket makes sure that duplicate socket
connections don't get established and now components and hooks that need
access to socket events, can get it directly by adding their own event
listeners to the socket element.