Refactor ServerView
Refactored ServerView handling of activeChannel. Before it used to be
stored in a redundant state variable - a leftover from back when
React Router wasn't being used to navigation. Now it's a regular
variable calculated at render-time. Changed the useEffect to reflect
that fact.

Also fixed a bug where the useEffect was navigating to an invalid path
when the "channelId" param was empty.