Fix useServerStore initializing with no servers
useServerStore was initializing server store with an empty array, only
later running a useEffect that set the value based on the useAuth user
object. This caused issues in some instances where the default content
for an empty serverList would flash during the initial render, before
the useEffect ran.

Now it correctly initializes serverList based on the user object if it's
available.