Refactor ServersMenu
Fixed the styling to allow scrolling when overflow happens. Made
onCloseMenu optional to allow for better UX when used inside of
ChannelsScreen. Changed from padding to margin around the server list
because it works consistently with overflow scroll. Made the border
display above "Add a server" conditional so that it doesn't appear
when the server list is empty. Also broke long className declarations
into mulitple lines.