Refactor ProtectedRoute component
Because of the asynchronous nature of logging in, ProtectedRoute would
briefly display Login page, before navigating to the authenticated
section of the app. This was confusing and caused an unpleasant, abrupt
flicker.

The component has been refactored to make use of new state inside
useAuth hook, which helps it understand when there is an in-flight
login request and allowing it to display a loading page while it is
happening, preventing the unpleasant flicker.