Remove useSocketEvents hook
The useSocketEvents hook and associated types have been removed.
This component was pointlessly duplicating and storing all events
coming from the socket and, in order to consume those copies, required
adding another instance of EventEmitter (on top of the one already
included with SocketIo). Now that all the components that used to rely
on its functionality have been refactored to use the events emitted
by the "socket" element directly, it's safe to remove it.