Fix useSwipe moving on single click
When used with the constrainToContainer option set to "false", the
hook would return a delta different than 0 when simply clicking without
moving the mouse, outside the container. It was caused by
handleSwipeStart calling "return" without updating the X,Y coordinates
when the click happens outside the swipe container. That led to the
inital X,Y coordinates and the handleSwipeEnd coordinates beind out of
sync, thus returning a non-0 delta.

A new isSwiping state has been introduced in order to keep the two
methods in-sync and fixing the issue.