Add swipe direction lock to useSwipeNavigation
Vertically scrolling the app's content was causing accidental drag
of the adjacent columns. After experimenting with other solutions such
as minimum drag distance, I finally settled on enforcing swipe direction
lock. The hook determines if the user is trying to scroll vertically or
slide horizontally and locks the rest of the swipe interaction in that
direction. It feels responsive to use and minimizes accidental
horizontal swipes.