Fix keyboard navigation in ChannelButton
The handleShortPressClick handler was expecting an ongoing mouse press
in order to navigate to the channel, which meant that keyboard
navigation wouldn't work. A separate handler was introduced and
extra checks were put in place to make sure the navigation doesn't
happen twice, since mouse events, unlike touch events, directly
trigger the button's onClick handler as well as the onCancel handler
inside useLongPress.