Refactor ServerSettings
Refactored ServerSettings component into a standalone page, instead of
an on-screen overlay. As a result, it has been changed to be able to
fetch server data by itself, as well as be able to receive it through
state passed in a React Router link. Added loading and error state
display for different stages of data fetching. Refactored
DeleteServerModal to make use of ModalContainer.