Fix multiple client messages not being displayed
When sending consecutive chat messages, only the first one would be
displayed, until the page was refreshed. The function which removes
duplicate messages was only using _id field for comparison, not
clientId. As a result subsequent client messages which share the
value of _id: undefined were being flagged as duplicates and removed.
Adding clientId to the list of unique keys fixes that issue.