Fix MessageInput value lost on breakpoint change
SwipeNavigation causes the component tree to get unmounted and replaced
with different components, meaning the current value of MessageInput
gets wiped every time a different breakpoint was triggered. This could
be highly frustrating for users on mobile devices or when resizing
browser window. By storing the message input value inside React context,
the editor is able to initialize with the previous value during
breakpoint changes, which fixes the problem.