Add message loading on scroll to ChatDisplay
Added LoadingBlock component which loads more messages when it's
scrolled into view in ChatDisplay and displays a loading spinner
while the new messages are loading. Updated useScrollPosition hook to
maintain the scroll position when more messages load and are added into
the container. Also added a loading indicator in ChatDisplay when the
message array is undefined.