Improve useScrollPosition scroll performance
Previous implementation was calling a rather heavy handler on every
scrollend event, which only got slower the more messages were loaded.
In order to prevent the lag caused by the handler, a debounce function
has been added. The event listener was moved from a "scrollend" to a
regular "scroll" event, because it allows for a smaller delay to be
set inside the debounce function, which helps avoid a situation where
the user navigates to a different channel before the debounced handler
executes.

isOffsetUpdateQueued state has been introduced to prevent a situation
where doing a constant scroll from the bottom of the container to the
point where infinite scroll loads more messages will cause the scroll
position to jump all the way to the bottom, since the debounced handler
hasn't updated the cursor value.