Add initial observe option to useResizeObserver
Added an optional to parameter to useResizeObserver that allows
declaring whether certain functionality should be performed when
ResizeObserver first observes the passed ref. Especially useful when
the callback reference changes (e.g. because of a dependency change)
which causes the useEffect to rerun and the same element gets observed
again.