Fix scroll bouncing to bottom on slight move up
When scrolling only slightly up from the bottom of the container the
scroll position would jump back to the bottom, but only the first time.
It was caused by scrollOffset changing value from undefined to 'bottom',
which triggered useResizeObserver callback.

Changing useResizeObserver options to not run the callback on initial
observe fixed the problem.