Fix small inconsistencies in ChatDisplay scroll
An incorrect min-height declaration in ChatMessage caused a layout
shift between the initial visual state and the state after effects
ran. This meant that automatic scrolling which happens during
that initial state was slightly off by the time the effect ran.

The component has been reworked to not have to rely on that min-height
declaration to avoid similar bugs in the future.