Fix unexpected scroll "jump" on new chat messages
Whenever a new message was received over websockets and added to
chat display, if scrollOffset wasn't set to "bottom", it would
unexpectedly shift by a seemingly random amount. The bug was caused
by an Effect inside useScrollPosition that was supposed to maintain
scroll position when older messages are fetched from the API and added
into chat container at the top of the container. The logic inside the
effect worked correctly when the messages were being added at the top,
but it caused a buggy scroll when the they were added at the bottom,
such as when receiving new messages over websockets.

The solution was to use a callback inside useChatMessages that would
notify the hook when changes to "messages" dependency were being caused
by older messages loading and allowing it to conditionally maintain
scroll position only in those instances.