Fix scroll being forced down on slight scroll up
When scrolling up inside ChatDisplay by a relatively small amount, the
scroll would keep getting pushed down to the very bottom of the
container. The problem was caused by the isOffsetUpdateQueued state
which changed value after every scroll event, causing a useEffect to
rerun and since the value of scrollOffset was still "bottom", the
effect would push the scroll position all the way down.

The state has been replaced with a useRef, which doesn't cause the
Effect to rerun and fixes the problem.