Add SettingsPage component
This component is based on the current ServerSettings component,
cleaning up some of its logic and making it generic enough to allow
it to be used with different kinds of settings pages (e.g. channel
settings).

Also changed the default blocking behavior of the onCloseSettings
prop. When that prop is passed a function that navigates to a different
URL it causes a situation where the Discard Changes modal is displayed
twice. Now you can specify on case-by-case basis if you want
onCloseSettings to be blocked or if you'd rather use the default
on-navigation blocking.