Add "complete account" redirect to ProtectedRoute
Refactored ProtectedRoute component to redirect the user to the
"CompleteAccountSetup" screen when they are authenticated, but their
account status isn't approved. The redirect passes a "returnTo" state
which can be used to return the user to the location they were trying
to access after they've finished setting up their account.