Fix ImageInput not updating displayed image
ImageInput worked fine when the initial currentImage value was empty,
but when used in settings forms that already had an intial currentImage
value, it failed to keep the displayed image in-sync with that value.

Introduced a new option prop that informs the component that the
provided currentImage value is a valid URL that can be displayed when
both presetImage and uploadedImage are empty.