Setup lazy loading on the /app route
Currently the /app route makes up for about 50% of the bundle size. In
order to improve app load times, that entire route has been switched
to lazy loading, allowing the bundler to split the code bundle.

An extra file re-exporting the components and container the route
relies on has been added. This way the entire route code is bundled
together, so that the browser only has to make a single request.