Fix server store not updating on server changes
When a server was updated, either by the user themselves or another
user, the changes would be reflected in ServerView, but the server
store and consequentially ServerSelectBar would retain the old value.

useServerStore has been changed so that it refetches server data on
every ServerUpdated socket event, preventing it from falling out of
sync.